[frontend] js to ts root file, knowledge force update hook added (#4352)
