+++ opencti-platform/opencti-front/src/private/components/arsenal/channels/Root.tsx
@@ -6,6 +6,7 @@ import Box from '@mui/material/Box';
 import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
 import useQueryLoading from 'src/utils/hooks/useQueryLoading';
+import useForceUpdate from '@components/common/bulk/useForceUpdate';
 import StixCoreObjectContentRoot from '../../common/stix_core_objects/StixCoreObjectContentRoot';
 import Channel from './Channel';
 import ChannelKnowledge from './ChannelKnowledge';
@@ -94,6 +95,8 @@ const RootChannel = ({ queryRef, channelId }: RootChannelProps) => {
     connectorsForExport,
     connectorsForImport,
   } = usePreloadedQuery<RootChannelQuery>(channelQuery, queryRef);
+
+  const { forceUpdate } = useForceUpdate();
 
   const paddingRight = getPaddingRight(location.pathname, channelId, '/dashboard/arsenal/channels');
   const link = `/dashboard/arsenal/channels/${channelId}/knowledge`;
@@ -210,7 +213,11 @@ const RootChannel = ({ queryRef, channelId }: RootChannelProps) => {
               />
               <Route
                 path="/knowledge/*"
-                element={<ChannelKnowledge channel={channel} />}
+                element={
+                  <div key={forceUpdate}>
+                    <ChannelKnowledge channel={channel} />
+                  </div>
+                }
               />
               <Route
                 path="/content/*"
