+++ opencti-platform/opencti-front/src/private/components/entities/events/Root.tsx
@@ -1,71 +1,65 @@
 import React, { useMemo, Suspense } from 'react';
-import { Route, Routes, Link, Navigate, useParams, useLocation } from 'react-router-dom';
+import { Route, Routes, Link, Navigate, useLocation, useParams } from 'react-router-dom';
 import { graphql, useSubscription, usePreloadedQuery, PreloadedQuery } from 'react-relay';
+import { GraphQLSubscriptionConfig } from 'relay-runtime';
 import Box from '@mui/material/Box';
 import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
+import { RootEventQuery } from '@components/entities/events/__generated__/RootEventQuery.graphql';
+import { RootEventsSubscription } from '@components/entities/events/__generated__/RootEventsSubscription.graphql';
 import useQueryLoading from 'src/utils/hooks/useQueryLoading';
-import { GraphQLSubscriptionConfig } from 'relay-runtime';
 import useForceUpdate from '@components/common/bulk/useForceUpdate';
 import StixCoreObjectContentRoot from '../../common/stix_core_objects/StixCoreObjectContentRoot';
-import StixCoreObjectSimulationResult from '../../common/stix_core_objects/StixCoreObjectSimulationResult';
-import Malware from './Malware';
-import MalwareKnowledge from './MalwareKnowledge';
+import Event from './Event';
+import EventKnowledge from './EventKnowledge';
 import StixDomainObjectHeader from '../../common/stix_domain_objects/StixDomainObjectHeader';
 import FileManager from '../../common/files/FileManager';
-import MalwarePopover from './MalwarePopover';
+import EventPopover from './EventPopover';
 import Loader, { LoaderVariant } from '../../../../components/Loader';
 import StixCoreObjectHistory from '../../common/stix_core_objects/StixCoreObjectHistory';
 import StixCoreObjectOrStixCoreRelationshipContainers from '../../common/containers/StixCoreObjectOrStixCoreRelationshipContainers';
 import StixCoreObjectKnowledgeBar from '../../common/stix_core_objects/StixCoreObjectKnowledgeBar';
 import ErrorNotFound from '../../../../components/ErrorNotFound';
+import EntityStixSightingRelationships from '../../events/stix_sighting_relationships/EntityStixSightingRelationships';
 import { useFormatter } from '../../../../components/i18n';
 import Breadcrumbs from '../../../../components/Breadcrumbs';
 import { getCurrentTab, getPaddingRight } from '../../../../utils/utils';
-import { RootMalwareQuery } from './__generated__/RootMalwareQuery.graphql';
-import { RootMalwareSubscription } from './__generated__/RootMalwareSubscription.graphql';
-import useHelper from '../../../../utils/hooks/useHelper';
 import Security from '../../../../utils/Security';
 import { KNOWLEDGE_KNUPDATE } from '../../../../utils/hooks/useGranted';
-import MalwareEdition from './MalwareEdition';
-import EditEntityControlledDial from '../../../../components/EditEntityControlledDial';
+import EventEdition from './EventEdition';
 
 const subscription = graphql`
-  subscription RootMalwareSubscription($id: ID!) {
+  subscription RootEventsSubscription($id: ID!) {
     stixDomainObject(id: $id) {
-      ... on Malware {
-        ...Malware_malware
-        ...MalwareEditionContainer_malware
+      ... on Event {
+        ...Event_event
+        ...EventEditionContainer_event
       }
       ...FileImportViewer_entity
       ...FileExportViewer_entity
       ...FileExternalReferencesViewer_entity
       ...WorkbenchFileViewer_entity
-      ...PictureManagementViewer_entity
     }
   }
 `;
 
-const malwareQuery = graphql`
-  query RootMalwareQuery($id: String!) {
-    malware(id: $id) {
+const eventQuery = graphql`
+  query RootEventQuery($id: String!) {
+    event(id: $id) {
       id
-      standard_id
       entity_type
       name
       aliases
-      x_opencti_graph_data
       stixCoreObjectsDistribution(field: "entity_type", operation: count) {
         label
         value
       }
-      ...Malware_malware
-      ...MalwareKnowledge_malware
+      ...Event_event
+      ...EventKnowledge_event
       ...FileImportViewer_entity
       ...FileExportViewer_entity
       ...FileExternalReferencesViewer_entity
       ...WorkbenchFileViewer_entity
-      ...PictureManagementViewer_entity
       ...StixCoreObjectContent_stixCoreObject
     }
     connectorsForImport {
@@ -77,37 +71,34 @@ const malwareQuery = graphql`
   }
 `;
 
-type RootMalwareProps = {
-  malwareId: string;
-  queryRef: PreloadedQuery<RootMalwareQuery>;
+type RootEventProps = {
+  eventId: string;
+  queryRef: PreloadedQuery<RootEventQuery>;
 };
 
-const RootMalware = ({ queryRef, malwareId }: RootMalwareProps) => {
-  const subConfig = useMemo<GraphQLSubscriptionConfig<RootMalwareSubscription>>(() => ({
+const RootEvent = ({ eventId, queryRef }: RootEventProps) => {
+  const subConfig = useMemo<GraphQLSubscriptionConfig<RootEventsSubscription>>(() => ({
     subscription,
-    variables: { id: malwareId },
-  }), [malwareId]);
+    variables: { id: eventId },
+  }), [eventId]);
 
   const location = useLocation();
   const { t_i18n } = useFormatter();
-  useSubscription<RootMalwareSubscription>(subConfig);
-  const { isFeatureEnable } = useHelper();
-  const isFABReplaced = isFeatureEnable('FAB_REPLACEMENT');
+  useSubscription<RootEventsSubscription>(subConfig);
 
   const {
-    malware,
+    event,
     connectorsForExport,
     connectorsForImport,
-  } = usePreloadedQuery<RootMalwareQuery>(malwareQuery, queryRef);
+  } = usePreloadedQuery<RootEventQuery>(eventQuery, queryRef);
 
   const { forceUpdate } = useForceUpdate();
 
-  const isOverview = location.pathname === `/dashboard/arsenal/malwares/${malwareId}`;
-  const paddingRight = getPaddingRight(location.pathname, malwareId, '/dashboard/arsenal/malwares');
-  const link = `/dashboard/arsenal/malwares/${malwareId}/knowledge`;
+  const link = `/dashboard/entities/events/${eventId}/knowledge`;
+  const paddingRight = getPaddingRight(location.pathname, eventId, '/dashboard/entities/events');
   return (
     <>
-      {malware ? (
+      {event ? (
         <>
           <Routes>
             <Route
@@ -116,106 +107,99 @@ const RootMalware = ({ queryRef, malwareId }: RootMalwareProps) => {
                 <StixCoreObjectKnowledgeBar
                   stixCoreObjectLink={link}
                   availableSections={[
-                    'victimology',
+                    'locations',
                     'threats',
                     'threat_actors',
                     'intrusion_sets',
                     'campaigns',
                     'incidents',
-                    'variants',
-                    'tools',
+                    'malwares',
                     'attack_patterns',
-                    'vulnerabilities',
-                    'indicators',
+                    'tools',
                     'observables',
-                    'infrastructures',
-                    'sightings',
                   ]}
-                  stixCoreObjectsDistribution={malware.stixCoreObjectsDistribution}
+                  stixCoreObjectsDistribution={event.stixCoreObjectsDistribution}
                 />
               }
             />
           </Routes>
           <div style={{ paddingRight }}>
             <Breadcrumbs variant="object" elements={[
-              { label: t_i18n('Arsenal') },
-              { label: t_i18n('Malwares'), link: '/dashboard/arsenal/malwares' },
-              { label: malware.name, current: true },
+              { label: t_i18n('Entities') },
+              { label: t_i18n('Events'), link: '/dashboard/entities/events' },
+              { label: event.name, current: true },
             ]}
             />
             <StixDomainObjectHeader
-              entityType="Malware"
-              stixDomainObject={malware}
-              PopoverComponent={<MalwarePopover />}
-              EditComponent={isFABReplaced && (
+              entityType="Event"
+              stixDomainObject={event}
+              enableQuickSubscription={true}
+              PopoverComponent={<EventPopover />}
+              EditComponent={(
                 <Security needs={[KNOWLEDGE_KNUPDATE]}>
-                  <MalwareEdition
-                    malwareId={malware.id}
-                    controlledDial={EditEntityControlledDial}
-                  />
+                  <EventEdition eventId={event.id} />
                 </Security>
               )}
-              enableQuickSubscription={true}
             />
             <Box
               sx={{
                 borderBottom: 1,
                 borderColor: 'divider',
                 marginBottom: 3,
-                display: 'flex',
-                justifyContent: 'space-between',
-                alignItem: 'center',
               }}
             >
               <Tabs
-                value={getCurrentTab(location.pathname, malware.id, '/dashboard/arsenal/malwares')}
+                value={getCurrentTab(location.pathname, event.id, '/dashboard/entities/events')}
               >
                 <Tab
                   component={Link}
-                  to={`/dashboard/arsenal/malwares/${malware.id}`}
-                  value={`/dashboard/arsenal/malwares/${malware.id}`}
+                  to={`/dashboard/entities/events/${event.id}`}
+                  value={`/dashboard/entities/events/${event.id}`}
                   label={t_i18n('Overview')}
                 />
                 <Tab
                   component={Link}
-                  to={`/dashboard/arsenal/malwares/${malware.id}/knowledge/overview`}
-                  value={`/dashboard/arsenal/malwares/${malware.id}/knowledge`}
+                  to={`/dashboard/entities/events/${event.id}/knowledge/overview`}
+                  value={`/dashboard/entities/events/${event.id}/knowledge`}
                   label={t_i18n('Knowledge')}
                 />
                 <Tab
                   component={Link}
-                  to={`/dashboard/arsenal/malwares/${malware.id}/content`}
-                  value={`/dashboard/arsenal/malwares/${malware.id}/content`}
+                  to={`/dashboard/entities/events/${event.id}/content`}
+                  value={`/dashboard/entities/events/${event.id}/content`}
                   label={t_i18n('Content')}
                 />
                 <Tab
                   component={Link}
-                  to={`/dashboard/arsenal/malwares/${malware.id}/analyses`}
-                  value={`/dashboard/arsenal/malwares/${malware.id}/analyses`}
+                  to={`/dashboard/entities/events/${event.id}/analyses`}
+                  value={`/dashboard/entities/events/${event.id}/analyses`}
                   label={t_i18n('Analyses')}
                 />
                 <Tab
                   component={Link}
-                  to={`/dashboard/arsenal/malwares/${malware.id}/files`}
-                  value={`/dashboard/arsenal/malwares/${malware.id}/files`}
+                  to={`/dashboard/entities/events/${event.id}/sightings`}
+                  value={`/dashboard/entities/events/${event.id}/sightings`}
+                  label={t_i18n('Sightings')}
+                />
+                <Tab
+                  component={Link}
+                  to={`/dashboard/entities/events/${event.id}/files`}
+                  value={`/dashboard/entities/events/${event.id}/files`}
                   label={t_i18n('Data')}
                 />
                 <Tab
                   component={Link}
-                  to={`/dashboard/arsenal/malwares/${malware.id}/history`}
-                  value={`/dashboard/arsenal/malwares/${malware.id}/history`}
+                  to={`/dashboard/entities/events/${event.id}/history`}
+                  value={`/dashboard/entities/events/${event.id}/history`}
                   label={t_i18n('History')}
                 />
               </Tabs>
-              {isOverview && (
-                <StixCoreObjectSimulationResult id={malware.id} type="threat" />
-              )}
             </Box>
             <Routes>
               <Route
                 path="/"
                 element={
-                  <Malware malwareData={malware} />
+                  <Event eventData={event} />
                 }
               />
               <Route
@@ -223,7 +207,7 @@ const RootMalware = ({ queryRef, malwareId }: RootMalwareProps) => {
                 element={
                   <Navigate
                     replace={true}
-                    to={`/dashboard/arsenal/malwares/${malwareId}/knowledge/overview`}
+                    to={`/dashboard/entities/events/${eventId}/knowledge/overview`}
                   />
                 }
               />
@@ -231,7 +215,7 @@ const RootMalware = ({ queryRef, malwareId }: RootMalwareProps) => {
                 path="/knowledge/*"
                 element={
                   <div key={forceUpdate}>
-                    <MalwareKnowledge malware={malware} />
+                    <EventKnowledge event={event} />
                   </div>
                 }
               />
@@ -239,7 +223,7 @@ const RootMalware = ({ queryRef, malwareId }: RootMalwareProps) => {
                 path="/content/*"
                 element={
                   <StixCoreObjectContentRoot
-                    stixCoreObject={malware}
+                    stixCoreObject={event}
                   />
                 }
               />
@@ -247,28 +231,39 @@ const RootMalware = ({ queryRef, malwareId }: RootMalwareProps) => {
                 path="/analyses"
                 element={
                   <StixCoreObjectOrStixCoreRelationshipContainers
-                    stixDomainObjectOrStixCoreRelationship={malware}
+                    stixDomainObjectOrStixCoreRelationship={event}
                   />
                 }
+              />
+              <Route
+                path="/sightings"
+                element={(
+                  <EntityStixSightingRelationships
+                    entityId={event.id}
+                    entityLink={link}
+                    noPadding={true}
+                    isTo={true}
+                  />
+                )}
               />
               <Route
                 path="/files"
-                element={
+                element={ (
                   <FileManager
-                    id={malwareId}
+                    id={eventId}
                     connectorsImport={connectorsForImport}
                     connectorsExport={connectorsForExport}
-                    entity={malware}
+                    entity={event}
                   />
-                }
+                )}
               />
               <Route
                 path="/history"
-                element={
+                element={ (
                   <StixCoreObjectHistory
-                    stixCoreObjectId={malwareId}
+                    stixCoreObjectId={eventId}
                   />
-                }
+                )}
               />
             </Routes>
           </div>
@@ -279,18 +274,17 @@ const RootMalware = ({ queryRef, malwareId }: RootMalwareProps) => {
     </>
   );
 };
-
 const Root = () => {
-  const { malwareId } = useParams() as { malwareId: string; };
-  const queryRef = useQueryLoading<RootMalwareQuery>(malwareQuery, {
-    id: malwareId,
+  const { eventId } = useParams() as { eventId: string; };
+  const queryRef = useQueryLoading<RootEventQuery>(eventQuery, {
+    id: eventId,
   });
 
   return (
     <>
       {queryRef && (
         <Suspense fallback={<Loader variant={LoaderVariant.container} />}>
-          <RootMalware queryRef={queryRef} malwareId={malwareId} />
+          <RootEvent eventId={eventId} queryRef={queryRef} />
         </Suspense>
       )}
     </>
