+++ opencti-platform/opencti-front/src/private/components/events/incidents/Root.tsx
@@ -12,6 +12,7 @@ import StixCoreObjectSimulationResult from '@components/common/stix_core_objects
 import StixCoreObjectContentRoot from '@components/common/stix_core_objects/StixCoreObjectContentRoot';
 import Security from 'src/utils/Security';
 import { KNOWLEDGE_KNUPDATE } from 'src/utils/hooks/useGranted';
+import useForceUpdate from '@components/common/bulk/useForceUpdate';
 import Incident from './Incident';
 import IncidentKnowledge from './IncidentKnowledge';
 import StixDomainObjectHeader from '../../common/stix_domain_objects/StixDomainObjectHeader';
@@ -88,6 +89,7 @@ const RootIncidentComponent = ({ queryRef }) => {
   const { t_i18n } = useFormatter();
   useSubscription(subConfig);
   const data = usePreloadedQuery(incidentQuery, queryRef);
+  const { forceUpdate } = useForceUpdate();
   const { incident, connectorsForImport, connectorsForExport } = data;
   const link = `/dashboard/events/incidents/${incidentId}/knowledge`;
   const isOverview = location.pathname === `/dashboard/events/incidents/${incident?.id}`;
@@ -213,7 +215,11 @@ const RootIncidentComponent = ({ queryRef }) => {
               />
               <Route
                 path="/knowledge/*"
-                element={<IncidentKnowledge incidentData={incident} />}
+                element={
+                  <div key={forceUpdate}>
+                    <IncidentKnowledge incidentData={incident} />
+                  </div>
+                }
               />
               <Route
                 path="/content/*"
