+++ opencti-platform/opencti-front/src/private/components/threats/campaigns/Root.tsx
@@ -7,6 +7,7 @@ import Tab from '@mui/material/Tab';
 import useQueryLoading from 'src/utils/hooks/useQueryLoading';
 import { GraphQLSubscriptionConfig } from 'relay-runtime';
 import { RootCampaignSubscription } from '@components/threats/campaigns/__generated__/RootCampaignSubscription.graphql';
+import useForceUpdate from '@components/common/bulk/useForceUpdate';
 import StixCoreObjectContentRoot from '../../common/stix_core_objects/StixCoreObjectContentRoot';
 import StixCoreObjectSimulationResult from '../../common/stix_core_objects/StixCoreObjectSimulationResult';
 import Campaign from './Campaign';
@@ -95,6 +96,8 @@ const RootCampaign = ({ campaignId, queryRef }: RootCampaignProps) => {
     connectorsForExport,
     connectorsForImport,
   } = usePreloadedQuery<RootCampaignQuery>(campaignQuery, queryRef);
+
+  const { forceUpdate } = useForceUpdate();
 
   const link = `/dashboard/threats/campaigns/${campaignId}/knowledge`;
   const isOverview = location.pathname === `/dashboard/threats/campaigns/${campaignId}`;
@@ -217,7 +220,11 @@ const RootCampaign = ({ campaignId, queryRef }: RootCampaignProps) => {
               />
               <Route
                 path="/knowledge/*"
-                element={<CampaignKnowledge campaign={campaign} />}
+                element={
+                  <div key={forceUpdate}>
+                    <CampaignKnowledge campaign={campaign} />
+                  </div>
+                }
               />
               <Route
                 path="/content/*"
